requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
sqlalchemy>=2.0.0
pydantic>=2.4.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
pytest>=7.4.0
httpx>=0.25.0
bcrypt>=4.0.1
alembic>=1.12.0